Fix and clean up PR CI (#740)
* Fixing the examples and test modules

After the migration out from GoogleContainerTools to kptdev, the v.0.1.0 pointer is not valid anymore in the examples and test repositories.


* Remove typescript from CI


* Remove some other files


* Fix CI


* Fix golangci-lint version


* An other try at fixing golangci


* Moving golangci to a github action


* Revert "An other try at fixing golangci"

This reverts commit 1038435a4d45aa4447cd326e2ac74ad21ea8c604.


* Revert "Moving golangci to a github action"

This reverts commit 016ce6b9c6a009ff11e6de40bcdcec16c22342ea.


* Fix checkout path to new repo url


* Fix style-doc reference in get-started project


* Bumping go version to 1.24.3 to be in line with the kptdev/kpt


* Add back license to golangci file


* Moving from kptdev/kpt from v1.0.0-beta.57 to v1.0.0-beta.58


---------

+++ go/kfn/commands/build_test.go
@@ -30,8 +30,9 @@ func TestBuild(t *testing.T) {
 		// expected args is key, and expected return is value
 		cmdExpected []string
 		// expected env var is key, and existence is value
-		lookPathExpected map[string]bool
-		expectedError    string
+		lookPathExpected   map[string]bool
+		expectedError      string
+		dockerfileExpected bool
 	}{
 		"default build is ko, ko already exists": {
 			args: []string{""},
@@ -41,6 +42,7 @@ func TestBuild(t *testing.T) {
 			lookPathExpected: map[string]bool{
 				"ko": true,
 			},
+			dockerfileExpected: false,
 		},
 		"default build is ko, ko not exists": {
 			args: []string{""},
@@ -51,6 +53,7 @@ func TestBuild(t *testing.T) {
 			lookPathExpected: map[string]bool{
 				"ko": false,
 			},
+			dockerfileExpected: false,
 		},
 		"ko as builder, specify repo": {
 			args: []string{"--repo=gcr.io/test"},
@@ -60,6 +63,7 @@ func TestBuild(t *testing.T) {
 			lookPathExpected: map[string]bool{
 				"ko": true,
 			},
+			dockerfileExpected: false,
 		},
 		"ko as builder, specify tag": {
 			args: []string{"--repo=gcr.io/test", "--tag=v1"},
@@ -69,13 +73,15 @@ func TestBuild(t *testing.T) {
 			lookPathExpected: map[string]bool{
 				"ko": true,
 			},
+			dockerfileExpected: false,
 		},
 		"docker as builder, docker not exists": {
 			args: []string{"--builder=docker"},
 			lookPathExpected: map[string]bool{
 				"docker": false,
 			},
-			expectedError: "kfn requires that `docker` is installed and on the PATH",
+			dockerfileExpected: false,
+			expectedError:      "kfn requires that `docker` is installed and on the PATH",
 		},
 		"docker as builder, docker exists": {
 			args: []string{"--builder=docker"},
@@ -85,6 +91,7 @@ func TestBuild(t *testing.T) {
 			lookPathExpected: map[string]bool{
 				"docker": true,
 			},
+			dockerfileExpected: true,
 		},
 		"docker as builder, specify dockerfile": {
 			args: []string{"--builder=docker", "--dockerfile=tmp/Dockerfile"},
@@ -94,6 +101,7 @@ func TestBuild(t *testing.T) {
 			lookPathExpected: map[string]bool{
 				"docker": true,
 			},
+			dockerfileExpected: true,
 		},
 		"docker as builder, specify image": {
 			args: []string{"--builder=docker", "--image=dockertest:latest", "--dockerfile=tmp/Dockerfile"},
@@ -103,27 +111,34 @@ func TestBuild(t *testing.T) {
 			lookPathExpected: map[string]bool{
 				"docker": true,
 			},
+			dockerfileExpected: true,
 		},
 	}
 	for name, test := range testcases {
-		r := NewBuildRunner(context.TODO())
-		execCmdFn = func(envs []string, name string, args ...string) error {
-			fakeExecCmd(t, test.cmdExpected, envs, name, args...)
-			return nil
-		}
-		execLookPathFn = func(file string) (string, error) {
-			return "", fakeExecLookPath(t, test.lookPathExpected, file)
-		}
-		r.Command.SetArgs(test.args)
-		err := r.Command.Execute()
-		if test.expectedError == "" {
-			if err != nil {
-				t.Errorf("%v failed. got error: %v", name, err)
+		t.Run(name, func(t *testing.T) {
+
+			r := NewBuildRunner(context.TODO())
+			execCmdFn = func(envs []string, name string, args ...string) error {
+				fakeExecCmd(t, test.cmdExpected, envs, name, args...)
+				return nil
 			}
-		} else {
-			assert.EqualError(t, err, test.expectedError)
-		}
-		os.Remove("Dockerfile")
+			execLookPathFn = func(file string) (string, error) {
+				return "", fakeExecLookPath(t, test.lookPathExpected, file)
+			}
+			r.Command.SetArgs(test.args)
+			err := r.Command.Execute()
+			if test.expectedError == "" {
+				if err != nil {
+					t.Errorf("%v failed. got error: %v", name, err)
+				}
+			} else {
+				assert.EqualError(t, err, test.expectedError)
+			}
+			if test.dockerfileExpected {
+				err = os.Remove("Dockerfile")
+				assert.NoError(t, err)
+			}
+		})
 	}
 }
